Render a valid JSON payload
The JSON payload was not valid previously. If we where triggered
on a non-pull-request event the pull request number would be empty
and thus the value for that key was empty. Also we where missing
a colon inside the JSON payload. To add insult to injury we where
also not rendering the BASH_ENV variables correctly since we where
single quoting and we where also not sourcing the BASH_ENV environment
variable before the second build step.

Hopefully with these fixes we should render a proper payload to send
to LogDNA.