CircleCI: correct pipeline
When calling orb-tools/publish apparently the workspace is not
attached by default, we need to do this as the packed orb YAML
is stored inside it. This is a strange default as for the
orb-tools/publish-dev command this is defaulted to true.

Also correct the name and the tag we wish to attach to the
released orb.